Move usePersonaCosmetics + add name color lightness correction
